# src/back/Imolt.Api/contracts/openapi.yaml
...
openapi: 3.0.3
info:
  title: ИМОЛТ — расчёт вывоза и утилизации отходов
  version: 1.2.0
  description: |
    Договор расчётной части сервиса ИМОЛТ: расчёт стоимости вывоза и
    утилизации строительных отходов по Москве и Московской области, подбор
    полигонов, маршрут, коммерческое предложение и заявка.

    **Как читать.** Операции сгруппированы по трём предметным областям
    расчётной части — «Справочники», «Расчёт», «Сделка». Границы областей
    объявлены записью ADR-0001, и группировка здесь — их внешняя проекция:
    по метке `x-область` видно, какая часть системы отвечает за операцию.

    **Что уже работает.** Отвечают только `/health` и `/ready`. Остальные
    операции помечены `x-состояние: объявлено`: форма запроса и ответа
    зафиксирована, сервер их пока не обслуживает. Мок поднимается из этого же
    файла, примеры ответов взяты из раздела 7 запроса на дизайн и согласованы
    между собой (адрес «Москва, ул. Годовикова, 9», дата актуальности
    17.09.2026).

    **Чего в договоре нет.** Поведение, не определённое заказчиком, не
    додумано: такие места помечены `x-вопросы` со ссылкой на открытый вопрос.
    Отложенные требования (R-047, R-062 — R-066) в версию 1 не входят.

    **Соглашения.**

    - Пути и имена полей — латиницей, тексты и описания — по-русски;
      идентификаторы кода остаются латиницей по правилам проекта.
    - Денежная сумма передаётся строкой с двумя знаками после точки: двоичная
...
      в макете (R-059).
  contact:
...
  license:
...
servers:
...
security:
  - {}
  - bearerAuth: []
paths:
...
  /v1/document-service-orders:
    post:
      operationId: createDocumentServiceOrder
      tags: [Сделка]
      summary: Заказать услугу по документации
      x-область: сделка
      x-требования: [R-009, R-052, R-054]
      x-вопросы: [Q-005, Q-011]
      x-экран: [Э-10]
      x-состояние: реализовано
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentServiceOrderInput' }
            example:
              serviceId: ossig-mo
              objectAddress: г Москва, ул Годовикова, д 9
              comment: снос ангара, начало работ 01.10.2026
              personalDataConsent: true
      responses:
        '201':
          description: заказ принят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentServiceOrder' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
  /v1/subscription-requests:
    post:
      operationId: createSubscriptionRequest
      tags: [Сделка]
      summary: Оставить заявку на подписку
      description: |
        Оплата вне сервиса: операция фиксирует намерение и переводит подписку
        в состояние «ожидает подтверждения». Тарифы и пробный период
        заказчиком не названы (Q-004), поэтому цена в договоре не объявляется.
      security:
        - bearerAuth: []
      x-область: сделка
      x-требования: [R-008, R-049]
      x-вопросы: [Q-004, Q-011]
      x-экран: [Э-09]
      x-состояние: реализовано
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriptionRequestInput' }
            example:
              role: carrier
              companyName: ООО «Перевозчик»
              inn: '7701234567'
              registeredInAisOssig: true
      responses:
        '201':
          description: заявка на подписку принята
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubscriptionRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
  /v1/auth/sessions:
    post:
      operationId: createSession
      tags: [Доступ]
      summary: Обменять параметры запуска MAX на сессию
      description: |
        Личность даёт платформа: мини-приложение передаёт сюда строку
        стартовых параметров `initData`, сервер проверяет её подпись ключом
        бота и в ответ выдаёт маркер доступа для заголовка
        `Authorization: Bearer` (ADR-0006). Пароль, телефон и код
        подтверждения здесь не передаются: собственный вход отложен
        требованием R-066 и в версию 1 не входит.

        Подпись проверяется на сервере: ключ бота наружу не выходит (R-056).
        Срок жизни маркера объявлен полем: клиент не угадывает его по опыту.
      security: []
      x-область: сделка
      x-требования: [R-049, R-050, R-051, R-054]
      x-вопросы: [Q-011]
      x-экран: [Э-09]
      x-состояние: реализовано
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionRequest' }
            example:
              initData: 'auth_date=1758067200&query_id=AAH…&user=%7B%22id%22%3A812345%7D&hash=0f3c…'
              personalDataConsent: true
      responses:
        '201':
          description: сессия создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
              example:
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.пример.пример
                expiresIn: 86400
                profile:
                  id: 9d3a7c10-52f4-4a1b-9ad8-0f6c2b5e41aa
                  maxUserId: '812345'
                  displayName: Иван
                  role: carrier
                  companyName: ООО «Перевозчик»
                  subscription: { state: pending }
        '401':
          description: подпись параметров запуска не сошлась или они устарели
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/profile:
    get:
      operationId: getProfile
      tags: [Доступ]
      summary: Профиль и состояние подписки
      description: |
        Кабинет показывает по этой операции пометку «Подписка ожидает
        подтверждения». Признак регистрации транспорта в АИС ОССиГ хранится
        в профиле перевозчика (R-051).
      security:
        - bearerAuth: []
      x-область: сделка
      x-требования: [R-049, R-050, R-051]
      x-вопросы: [Q-011]
      x-экран: [Э-10]
      x-состояние: реализовано
      responses:
        '200':
          description: профиль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Маркер сессии из операции `createSession`. Расчёт и коммерческое
        предложение задуманы доступными без входа; окончательный перечень
        закрытых операций — открытый вопрос Q-011.
...
  responses:
...
    NotFound:
      description: запись не найдена
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:not-found'
            title: Запись не найдена
            status: 404
    Unauthorized:
      description: операция требует входа
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:authentication-required'
            title: Нужен вход
            status: 401
...
    UnprocessableContent:
      description: запрос разобран, но нарушает правило предметной области
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:allocation-mismatch'
            title: Распределение не сходится с объёмом
            status: 422
            detail: По группе «Лом бетона» распределено 18 т из 20 т
            errors:
              - field: entries[1].quantity.value
                message: увеличьте на 2 т либо уменьшите объём группы
...
    TooManyRequests:
      description: обращения идут чаще допустимого
      headers:
        Retry-After:
          description: через сколько секунд повторить
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'urn:imolt:problem:too-many-requests'
            title: Слишком частые обращения
            status: 429
...
  schemas:
...
    Problem:
      type: object
      description: |
        Описание ошибки по RFC 9457. Поле `type` — устойчивый код причины:
        интерфейс ветвится по нему, а не по тексту заголовка.
      required: [type, title, status]
      properties:
        type: { type: string, example: 'urn:imolt:problem:validation' }
        title: { type: string, example: Запрос не прошёл проверку }
        status: { type: integer, example: 400 }
        detail: { type: string }
        instance: { type: string, description: 'путь запроса, на котором возникла ошибка' }
        errors:
          type: array
          description: разбор по полям — для форм
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    DocumentServiceOrderInput:
      type: object
      required: [serviceId, objectAddress, personalDataConsent]
      properties:
        serviceId: { type: string }
        objectAddress: { type: string }
        comment: { type: string }
        personalDataConsent: { type: boolean }
      additionalProperties: false
    DocumentServiceOrder:
      type: object
      required: [id, serviceId, createdAt, state]
      properties:
        id: { type: string }
        serviceId: { type: string }
        createdAt: { type: string, format: date-time }
        state: { type: string, enum: [accepted] }
        message: { type: string }
    SubscriberRole:
      type: string
      description: роль подписчика — перевозчик либо демонтажная компания
      enum: [carrier, demolitionCompany]
    SubscriptionState:
      type: object
      required: [state]
      properties:
        state:
          type: string
          description: |
            `none` — подписки нет; `pending` — заявка оставлена, оплата идёт
            вне сервиса; `active` — подписка подтверждена.
          enum: [none, pending, active]
        activeUntil: { type: string, format: date, nullable: true }
    SubscriptionRequestInput:
      type: object
      required: [role, companyName, inn]
      properties:
        role: { $ref: '#/components/schemas/SubscriberRole' }
        companyName: { type: string }
        inn: { type: string, pattern: '^[0-9]{10}$|^[0-9]{12}$' }
        registeredInAisOssig:
          type: boolean
          description: транспорт зарегистрирован в АИС ОССиГ (R-051)
      additionalProperties: false
    SubscriptionRequest:
      type: object
      required: [id, createdAt, subscription]
      properties:
        id: { type: string }
        createdAt: { type: string, format: date-time }
        subscription: { $ref: '#/components/schemas/SubscriptionState' }
        message: { type: string }
    SessionRequest:
      type: object
      required: [initData, personalDataConsent]
      properties:
        initData:
          type: string
          maxLength: 4096
          description: |
            Строка стартовых параметров мини-приложения MAX как есть, без
            разбора на стороне клиента: сервер сам разбирает её и проверяет
            подпись поля `hash`. Разобранный клиентом объект не принимается —
            проверять подпись можно только по исходной строке.
        personalDataConsent:
          type: boolean
          description: явное согласие на обработку персональных данных (R-054)
      additionalProperties: false
    Profile:
      type: object
      required: [id, maxUserId, subscription]
      properties:
        id: { type: string }
        maxUserId:
          type: string
          description: учётная запись платформы MAX, по которой опознан пользователь
        displayName: { type: string, nullable: true }
        role:
          allOf: [{ $ref: '#/components/schemas/SubscriberRole' }]
          nullable: true
        companyName: { type: string, nullable: true }
        inn: { type: string, nullable: true }
        registeredInAisOssig: { type: boolean, nullable: true }
        subscription: { $ref: '#/components/schemas/SubscriptionState' }
    Session:
      type: object
      required: [accessToken, expiresIn, profile]
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer, description: сколько секунд маркер действует }
        profile: { $ref: '#/components/schemas/Profile' }
...